Move AHFE refit PoC to attic
