+++ attic/training/refit_freesolv.py
